[#7] Refactored dashboard page's folder structure and incremented dashboard menu
