+++ src/components/Dashboard/Menu/styles.css
@@ -1,0 +1,3 @@
+.header {
+    margin-bottom: 5vh;
+}
