--- src/components/DashboardMenu/styles.css
@@ -1,9 +1,0 @@
-.custom-card-column {
-    column-count: 2;
-    margin: auto 4vw;
-}
-
-.custom-card {
-    /*max-width: calc(35vw);*/
-    /*max-height: calc(25vh);*/
-}
