--- src/pages/DashboardPage/styles.css
@@ -1,3 +1,0 @@
-.header {
-    margin-top: 5vh;
-}
